Improve hash function for CanonicalizedPath
For files that exist (which is a very common case), CanonicalizedPath::path
is "". This means that every CanonicalizedPath that points to a file that
exists had the same hash. This is obviously a pretty bad idea.

xoring in the inode number of the file improved performance of hash table
insertion by over 100x. I was testing with Spotify's client core library,
where the hash table insertion of Paths used to be a sore point. It went
from 1.2 seconds to 9 milliseconds in my tests.