Remove the commands tool
The idea of printing a list of commands to run is nice but it is overly simplistic
I think. I want to be able to do things like isolating the commands from environment
variables, doing linting and so on. Reducing the build system to just running a set
of commands like this is not helping there.