Fix race condition and stop implicitly starting the shk-trace server
The race condition is that it would start tracing the first thread it would
see for the process to be traced, but if tracing happened quickly enough, it
could start tracing the main thread of the application, which would be waiting
for tracing to finish, causing a deadlock.

This commit also removes shk-trace's ability to implicitly start the
tracing server if there is none. This would be helpful if shk-trace
was to be used manually, but it's not. We could just as well spawn it
in shk and let it live for as long as that process.

This makes it easier to debug because the server is not a daemon with
no stdout, and it also (along with suicide-when-orphaned) erases the risk
that the server process lives forever after shk-trace is no longer used.