Store only a bool instead of Fingerprint::timestamp in the log
The fact that it was a timestamp like this caused issues because there
were lots and lots of fingerprints that were the same except for the
timestamp, causing them to be duplicated a lot and eventually a very
big invocation log and slow no-op builds because of that.

The timestamp turns out to not actually be needed. All we need to know
is if the fingerprint is potentially racily clean. So that is what
we store now instead.

I did a quick test building Shuriken and found that the invocation log
became 85% smaller after this change.

This is a breaking change to the invocation log format.