Use separate sequences for fingerprint and path ids in invocation log
This allows the parsing code to do less mapping work, and to allocate
less memory for the path id to string "map" vector, which should save
a little bit of time.

This actually saves a rather substantial amount of time. On my largish
test project, parsing the ~80MB log takes below 400ms, down from above
500ms, maybe a 140ms improvement.

The main reason I'm doing this is that it will make it easier to change
the code to do less copying in the performance critical parsing phase.

This is a breaking change to the invocation log format.