Don't reject fingerprints just because they are to special files
There are cases where special block devices like /dev/autofs_nowait
or others sneak into inputs of commands. Instead of failing the
build, let's allow it. If programs use non-files as input it's
their problem to make sure to behave as if files and command line
params were the only inputs. (This is consistent with ignoring
network access for example.)