Don't invoke sudo in the default "all" target
This causes various issues, including making Shuriken not able
to build itself because sudo is invoked from a non-terminal target,
which causes Shuriken to try to fingerprint /sbin/sudo which it
can't because it can't read it.

It's not really right to call sudo, which might require user
interaction, from a non-console pool build step anyway.