Use no_direct_dependencies_built in canSkipBuildCommand
This is when all the work related to this flag actually gets to
have an effect: Now, in rather many cases when canSkipBuildCommand
would previously have had to stat files, potentially many files
(hundreds or even thousands is common with C++ if there are many
include search paths), it can instantly skip and make the decision
that nothing needs to be done.

This optimization applies when there are overspecified dependencies.
For example, CMake conservatively and usually unnecessarily adds an
order only dependency from all compile steps for a static library to
all other static library that the library depends on. For example, all
test .cpp files for a given library will have that library as an
order-only dependency. Before this patch, if any change was made to
the library implementation, canSkipBuildCommand would stat all input
files to all the test files for that library. After this patch it can
instantly decide that nothing needs to be done with no syscalls.