Be more careful about file modes when fingerprinting
This is preparation for being able to have a distributed cache. When
Shuriken needs to be able to insert the output files of the build steps,
it needs to keep track of file modes in a more thought-out way so it
actually makes sense between machines.

The behavior that this patch introduces is modelled after git.