Require that Subscriptions are default constructible
This is a step towards making destroying a Subscription mean cancellation.
When that is the case, cancellation can be defined as move assigning a
Subscription with a default constructed value.

This can also help some operators to avoid optional or unique_ptr just
because they need to be able to have an uninitialized subscription at
some times.